requirements.txt
# Hospital Management SaaS — Python dependencies
# Generated after audit fixes — February 2026

# Core framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
email-validator>=2.2.0

# Database
sqlalchemy>=2.0.29
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9

# 2FA — FIX: Added for TOTP/Google Authenticator support
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# Payment Gateways — FIX: Real SDK integration
razorpay>=1.4.2
stripe>=9.0.0

# Email
aiosmtplib>=3.0.1

# Notifications
firebase-admin>=6.5.0

# PDF generation
reportlab>=4.2.0
#weasyprint>=62.3

# Background tasks & scheduling
# (asyncio-based — no Celery required for basic scheduling)
redis>=5.0.4
aiofiles>=24.1.0

# HTTP client
httpx>=0.27.0

# Utilities
python-dateutil>=2.9.0
python-dotenv>=1.0.1
pillow>=10.3.0

# Dev/Test
pytest>=8.1.0
pytest-asyncio>=0.23.6
httpx>=0.27.0
